separate testing so each model has its own test suite
